Develop

+++ src/pages/Avatars/index.jsx
@@ -6,12 +6,13 @@ import { Modal } from '../../components/Modal/Modal';
 import { ImageUploader } from '../../components/ImageUploader/ImageUploader';
 import { ref, uploadBytesResumable, getDownloadURL, deleteObject} from "firebase/storage";
 import storage from '../../hooks/useFirebase';
-import { useAuth } from '../../hooks/useAuth';
+import { useAuth, useValidateToken } from '../../hooks/useAuth';
 import { useEffect, useState } from 'react'
 
 export const Avatars = () => {
     const [showEdit, setShowEdit] = useState(false);
     const [showDelete, setShowDelete] = useState(false);
+    const [showAdd, setShowAdd] = useState(false);
     const [avatars, setAvatars] = useState([]);
     const [selectedAvatar, setSelectedAvatar] = useState({});
     
@@ -24,25 +25,31 @@ export const Avatars = () => {
 
     useEffect(() => {
         getAvatars()
-    }, [])
+    }, [avatars])
 
     // For showing the edit modal
-    const handleEditClick = (identifier) => {
+    const handleEditClick = (avatar) => {
         setShowEdit(!showEdit)
-        setSelectedAvatar(identifier)
+        setSelectedAvatar(avatar)
     }
 
     // For showing the delete modal
-    const handleDeleteClick = () => {
+    const handleDeleteClick = (avatar) => {
         setShowDelete(!showDelete)
+        setSelectedAvatar(avatar)
     }
 
     // For adding a new Avatar
     const handleAddClick = () => {
-        alert("Agregar")
+        setShowAdd(!showAdd)
     }
 
     const updateAvatar = async (imageUrl) => {
+        if (!useValidateToken()) {
+            alert("Token expired");
+            return;
+        }
+
         // Delete the old image from firebase
         deleteFromFirebase(selectedAvatar.picture)
 
@@ -53,7 +60,7 @@ export const Avatars = () => {
         
         // Update the avatar with the new image
         try {
-            const response = await fetch(`https://api.mingo.studio/api/avatar/${selectedAvatar._id}`, {
+            fetch(`https://api.mingo.studio/api/avatar/${selectedAvatar._id}`, {
                 method: "PUT",
                 crossDomain:true,
                 headers: {
@@ -66,20 +73,32 @@ export const Avatars = () => {
                     "picture": `${imageUrl}`
                 })  
             })
-            .then((response) => response.json());
-
-            getAvatars()
+            .then((response) => {
+                response.json()
+                setShowEdit(false)
+            })
         } catch(error) {
             console.log(error)
         }
     }
 
-    const deleteAvatar = async (imageUrl) => {
-        // Delete the old image from firebase
-        deleteFromFirebase(selectedAvatar.picture)
+    const deleteAvatar = async () => {
+        if (!useValidateToken()) {
+            alert("Token expired");
+            return;
+        }
 
+        // Delete the old image from firebase
+        deleteFromFirebase(selectedAvatar?.picture)
+
+        // Get token from localstorage
+        const dataStorage = useAuth()
+        const token = dataStorage.token
+
+
+        // Delete Avatar
         try {
-            const response = await fetch(`https://api.mingo.studio/api/avatar/${selectedAvatar._id}`, {
+            fetch(`https://api.mingo.studio/api/avatar/${selectedAvatar._id}`, {
                 method: "DELETE",
                 crossDomain:true,
                 headers: {
@@ -87,14 +106,49 @@ export const Avatars = () => {
                     Accept: "application/json",
                     "Access-Control-Allow-Origin": "*", 
                     "Authorization": `bearer ${token}`
-                },
+                }
+            })
+            .then((response) => {
+                response.json()
+                setShowDelete(false)
+            })
+
+        } catch(error) {
+            console.log(error)
+        }
+    }
+
+    // Add Avatar
+    const addAvatar = (imageUrl) => {
+        if (!useValidateToken()) {
+            alert("Token expired");
+            return;
+        }
+
+        // Get token from localstorage
+        const dataStorage = useAuth()
+        const token = dataStorage.token
+
+        try {
+            fetch("https://api.mingo.studio/api/avatar", {
+                method: "POST",
+                crossDomain:true,
+                headers: {
+                    "Content-Type" : "application/json",
+                    Accept: "application/json",
+                    "Access-Control-Allow-Origin": "*", 
+                    "Authorization": `bearer ${token}`
+                }, 
                 body: JSON.stringify({
                     "picture": `${imageUrl}`
-                })  
+                })
             })
-            .then((response) => response.json());
+            .then((response) => {
+                response.json()
+                setShowDelete(false)
+            })
+            .cath((error) => console.log(error));
 
-            getAvatars()
         } catch(error) {
             console.log(error)
         }
@@ -116,17 +170,25 @@ export const Avatars = () => {
     // Save in Firebase
     const uploadToFirebase = async (images) => {
         // images is an array which has data_url and the file
-        if(!images) {
+        if(!images || images.length <= 0) {
             alert("Please select an image")
+            return;
         }
 
         // Gets the storage location to save the image
-        const storageRef = await ref(storage, `/mingo-avatar-images/${images[0].file.name}`)
+        const storageRef = ref(storage, `/mingo-avatar-images/${images[0].file.name}`)
         // Uploads the image in the storage reference defined before
-        const uploadTask = uploadBytesResumable(storageRef, images[0].file)
+        uploadBytesResumable(storageRef, images[0].file)
         .then(() => {
             getDownloadURL(storageRef).then((url) => {
-                updateAvatar(url)
+                if (showEdit) {
+                    updateAvatar(url)
+                    setShowEdit(false)
+                }
+                else {
+                    addAvatar(url)
+                    setShowAdd(false)
+                }
             });
         });
     }
@@ -144,15 +206,32 @@ export const Avatars = () => {
                     <AiOutlinePlus />
                 </button>
              </IconContext.Provider>
-
-             <Modal handleClickOpen={handleEditClick} show={showEdit}>
-                <span className={ [classes["Span"], classes["Title"]].join(" ") }>Edit</span>
-                <ImageUploader handleSaveClick={uploadToFirebase} number = {1}/>
-            </Modal>
-
-            <Modal handleClickOpen={handleDeleteClick} show={showDelete}>
-                <span className={ [classes["Span"], classes["Title"]].join(" ") }>Delete</span>
-            </Modal>
+             {
+                showEdit && selectedAvatar.picture ?
+                <Modal handleClickOpen={handleEditClick} show={showEdit}>
+                    <span className={ [classes["Span"], classes["Title"]].join(" ") }>Edit</span>
+                    <ImageUploader item={selectedAvatar} handleSaveClick={uploadToFirebase} number = {1}/>
+                </Modal> :
+                <></>
+             }
+             {
+                showDelete ?
+                <Modal handleClickOpen={handleDeleteClick} show={showDelete} w="27.5rem" h="15rem">
+                    <span className={ [classes["Span"], classes["Title"]].join(" ") }>Delete</span>
+                    <div className={ classes["ButtonContainer"] }>
+                        <button className={ classes['SaveButton'] } onClick={deleteAvatar}>Delete</button>
+                    </div>
+                </Modal> :
+                <></>
+             }
+             {
+                showAdd ? 
+                <Modal handleClickOpen={handleAddClick} show={showAdd}>
+                    <span className={ [classes["Span"], classes["Title"]].join(" ") }>Add</span>
+                    <ImageUploader handleSaveClick={uploadToFirebase} number = {1}/>
+                </Modal> :
+                <></>
+             }
         </div>
     )
 }
